resync with current ircd-hybrid-6 tree
git-svn-id: http://svn.ratbox.org/svnroot/trunk@511 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ include/res.h
@@ -45,7 +45,7 @@ extern unsigned long   cres_mem(struct Client* cptr);
  * name is modified only if there is enough space in the buffer to hold
  * the suffix
  */
-extern void add_local_domain(char* name, size_t len);
+extern void add_local_domain(char* name, int len);
 
 #endif /* INCLUDED_res_h */
